Add date formatting utils and clean up
