+++ shared/index.dart
@@ -2,3 +2,4 @@ export 'custom_fab.dart';
 export 'small_chip.dart';
 export 'form/standard/standard_form_field.dart';
 export 'form/standard/standard_form_section.dart';
+export 'utils/date_formatting.dart';
